pyproject.toml
[project]
name = "buybot"
version = "1.0.0"
description = "Умный помощник покупок на базе LLM и RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.3.0",
    "chromadb>=0.5.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "pandas>=2.0.0",
    "sentence-transformers>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-html>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short -m 'not slow'"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
